remove duplicate test and commented out sections
