+++ app/controllers/movies_controller.rb
@@ -7,7 +7,6 @@ class MoviesController < ApplicationController
     if params[:top_movies] == 'top rated'
       @facade = Top20Facade.new(params[:top_movies])
     elsif !params[:search].nil?
-      # search
       @facade = MovieSearchFacade.new(params[:search])
     end
   end
