+++ app/controllers/users_controller.rb
@@ -8,7 +8,6 @@ class UsersController < ApplicationController
 
   def show
     @facade = UserFacade.new(params[:id])
-    # @user = User.find(params[:id])
   end
 
   def create
