+++ app/services/movie_service.rb
@@ -8,7 +8,6 @@ class MovieService
   def conn
     Faraday.new(url: 'https://api.themoviedb.org') do |faraday|
       faraday.params['api_key'] = Rails.application.credentials.tmdb[:key]
-      # faraday.params['language'] = 'en-US'
     end
   end
