+++ spec/features/movies/index_spec.rb
@@ -9,6 +9,15 @@ RSpec.describe 'discover movies' do
     @user3 = User.create!(name: 'Sooyung', email: 'sooyung@turing.edu')
   end
 
+  it 'has a home link, discover movies title, search field, find top rated movies and find movies buttons' do
+    visit discover_user_path(@user1)
+    expect(page).to have_link('Home')
+    expect(page).to have_content('Discover Movies')
+    expect(page).to have_field('search')
+    expect(page).to have_button('Find Top Rated Movies')
+    expect(page).to have_button('Find Movies')
+  end
+
   it 'shows the top rated movies and their average rating' do
     json_response = File.read("spec/fixtures/top_rated.json")
     parsed = JSON.parse(json_response, symbolize_names: true)
@@ -41,8 +50,6 @@ RSpec.describe 'discover movies' do
   
   it 'returns objects that are searched for' do
     json_response = File.read('spec/fixtures/star_wars.json')
-    # parsed = JSON.parse(json_response, symbolize_names: true)
-    # movies = parsed[:results]
     stub_request(:get, "https://api.themoviedb.org/3/search/movie?api_key=#{Rails.application.credentials.tmdb[:key]}&query=Star%20Wars")
     .with(
       headers: {
