+++ spec/features/users/discover_spec.rb
@@ -10,8 +10,6 @@ RSpec.describe 'discover movies' do
   end
   it 'click find top rated movies' do
     json_response = File.read('spec/fixtures/top_rated.json')
-    # parsed = JSON.parse(json_response, symbolize_names: true)
-    # movies = parsed[:results]
     stub_request(:get, "https://api.themoviedb.org/3/movie/top_rated?api_key=#{Rails.application.credentials.tmdb[:key]}")
       .with(
         headers: {
@@ -27,8 +25,6 @@ RSpec.describe 'discover movies' do
   end
   it 'should have a search button' do
     json_response = File.read('spec/fixtures/empty_search.json')
-    # parsed = JSON.parse(json_response, symbolize_names: true)
-    # movies = parsed[:results]
     stub_request(:get, "https://api.themoviedb.org/3/search/movie?api_key=#{Rails.application.credentials.tmdb[:key]}&query=")
       .with(
         headers: {
@@ -45,8 +41,6 @@ RSpec.describe 'discover movies' do
   end
   it 'allows for a search' do
     json_response = File.read('spec/fixtures/star_wars.json')
-    # parsed = JSON.parse(json_response, symbolize_names: true)
-    # movies = parsed[:results]
     stub_request(:get, "https://api.themoviedb.org/3/search/movie?api_key=#{Rails.application.credentials.tmdb[:key]}&query=Star%20Wars")
       .with(
         headers: {
